FW-4394 Move context fetching to mixin
