+++ firstvoices/backend/views/base_views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 from rules.contrib.rest_framework import AutoPermissionViewSetMixin
 
 from backend import permissions
-from backend.models import Site
+from backend.models import Alphabet, Character, CharacterVariant, IgnoredCharacter, Site
 from backend.permissions import utils
 
 
@@ -71,3 +71,33 @@ class SiteContentViewSetMixin:
             return site
         else:
             raise PermissionDenied
+
+
+class DictionarySerializerContextMixin:
+    """
+    Adds context to a view which is passed to the dictionary serializer to remove the need for duplicate queries.
+    """
+
+    def get_serializer_context(self):
+        """
+        A helper function to gather additional model context which can be reused for multiple dictionary entries.
+        """
+
+        site = self.get_validated_site()[0]
+
+        context = super().get_serializer_context()
+
+        alphabet = Alphabet.objects.filter(site=site).first()
+        ignored_characters = IgnoredCharacter.objects.filter(site=site).values_list(
+            "title", flat=True
+        )
+        base_characters = Character.objects.filter(site=site).order_by("sort_order")
+        character_variants = CharacterVariant.objects.filter(site=site)
+        ignorable_characters = IgnoredCharacter.objects.filter(site=site)
+
+        context["alphabet"] = alphabet
+        context["ignored_characters"] = ignored_characters
+        context["base_characters"] = base_characters
+        context["character_variants"] = character_variants
+        context["ignorable_characters"] = ignorable_characters
+        return context
