+++ firstvoices/backend/views/word_of_the_day_views.py
@@ -5,7 +5,6 @@ from drf_spectacular.utils import OpenApiResponse, extend_schema, extend_schema_
 from rest_framework import mixins, viewsets
 from rest_framework.response import Response
 
-from backend.models import Alphabet, Character, CharacterVariant, IgnoredCharacter
 from backend.models.dictionary import (
     DictionaryEntry,
     TypeOfDictionaryEntry,
@@ -13,7 +12,11 @@ from backend.models.dictionary import (
 )
 from backend.permissions import utils
 from backend.serializers.word_of_the_day_serializers import WordOfTheDayListSerializer
-from backend.views.base_views import FVPermissionViewSetMixin, SiteContentViewSetMixin
+from backend.views.base_views import (
+    DictionarySerializerContextMixin,
+    FVPermissionViewSetMixin,
+    SiteContentViewSetMixin,
+)
 
 
 @extend_schema_view(
@@ -30,6 +33,7 @@ class WordOfTheDayView(
     FVPermissionViewSetMixin,
     SiteContentViewSetMixin,
     mixins.ListModelMixin,
+    DictionarySerializerContextMixin,
     viewsets.GenericViewSet,
 ):
     """
@@ -144,27 +148,3 @@ class WordOfTheDayView(
             queryset, many=True, context=self.get_serializer_context()
         )
         return Response(serializer.data)
-
-    def get_serializer_context(self):
-        """
-        A helper function to gather additional model context which can be reused for multiple dictionary entries.
-        """
-
-        site = self.get_validated_site()[0]
-
-        context = super().get_serializer_context()
-
-        alphabet = Alphabet.objects.filter(site=site).first()
-        ignored_characters = IgnoredCharacter.objects.filter(site=site).values_list(
-            "title", flat=True
-        )
-        base_characters = Character.objects.filter(site=site).order_by("sort_order")
-        character_variants = CharacterVariant.objects.filter(site=site)
-        ignorable_characters = IgnoredCharacter.objects.filter(site=site)
-
-        context["alphabet"] = alphabet
-        context["ignored_characters"] = ignored_characters
-        context["base_characters"] = base_characters
-        context["character_variants"] = character_variants
-        context["ignorable_characters"] = ignorable_characters
-        return context
